requirements.txt
pandas 
numpy
plotly
emoji 
transformers[torch]
torch_optimizer
Flask
requests
openpyxl
datasets
# datasets 
# evaluate 
# tensorflow
# python-dotenv